Move attach mode to the mlops
