+++ mlops/parallelm/mlops/__init__.py
@@ -14,6 +14,5 @@ symlink_name = "mlops"
 from .versions import Versions
 from .mlops import MLOps
 from .mlops import mlops
-from .mlops_devel import mlops_devel
 
 from .stats_category import StatCategory
